requirements.txt
botbuilder-core
botbuilder-integration-aiohttp
fastapi
python-dotenv
requests
uvicorn